Porting BUMsort benchmark
