+++ src/main/scala/leon/invariant/util/TreeUtil.scala
@@ -238,7 +238,8 @@ object ProgramUtil {
    */
   def userLevelFunctions(program: Program): Seq[FunDef] = {
     program.units.flatMap { u =>
-      u.definedFunctions.filter(fd => !fd.isTheoryOperation && (u.isMainUnit || !(fd.isLibrary || fd.isInvariant)))
+      u.definedFunctions.filter(fd => !fd.annotations.contains("theoryop") &&
+          (u.isMainUnit || !(fd.annotations.contains("library") || fd.isInvariant)))
     }
   }
