+++ src/main/scala/leon/laziness/FunctionsManager.scala
@@ -39,7 +39,7 @@ class FunctionsManager(p: Program) {
               args map rec(WithState())
             case f @ FunctionInvocation(_, Seq(CaseClass(_, Seq(invExpr)))) if isStarInvocation(f)(p) =>
               rec(Star())(invExpr)
-            case f @ FunctionInvocation(TypedFunDef(callee, _), args) if callee.isRealFunction =>
+            case f @ FunctionInvocation(TypedFunDef(callee, _), args) if !callee.canBeStrictField => // ignoring vals. Note: lazy vals will become memoized functions
               dg.addEdge(fd, callee, l)
               args map rec(l)
             case Ensuring(e, post) =>
